+++ dot/core/service.go
@@ -18,6 +18,8 @@ package core
 
 import (
 	"bytes"
+	"errors"
+	"fmt"
 	"math/big"
 	"sync"
 
@@ -27,7 +29,6 @@ import (
 	"github.com/ChainSafe/gossamer/lib/common"
 	"github.com/ChainSafe/gossamer/lib/common/optional"
 	"github.com/ChainSafe/gossamer/lib/crypto/sr25519"
-	"github.com/ChainSafe/gossamer/lib/database"
 	"github.com/ChainSafe/gossamer/lib/keystore"
 	"github.com/ChainSafe/gossamer/lib/runtime"
 	"github.com/ChainSafe/gossamer/lib/services"
@@ -95,7 +96,7 @@ type Config struct {
 // session, and network service.
 func NewService(cfg *Config) (*Service, error) {
 	if cfg.Keystore == nil {
-		return nil, ErrNilKeystore
+		return nil, fmt.Errorf("no keystore provided")
 	}
 
 	keys := cfg.Keystore.Sr25519Keypairs()
@@ -105,11 +106,11 @@ func NewService(cfg *Config) (*Service, error) {
 	}
 
 	if cfg.BlockState == nil {
-		return nil, ErrNilBlockState
+		return nil, fmt.Errorf("block state is nil")
 	}
 
 	if cfg.StorageState == nil {
-		return nil, ErrNilStorageState
+		return nil, fmt.Errorf("storage state is nil")
 	}
 
 	codeHash, err := cfg.StorageState.LoadCodeHash()
@@ -139,7 +140,7 @@ func NewService(cfg *Config) (*Service, error) {
 
 	if cfg.IsBabeAuthority {
 		if cfg.Keystore.NumSr25519Keys() == 0 {
-			return nil, ErrNoKeysProvided
+			return nil, fmt.Errorf("no keys provided for authority node")
 		}
 
 		epochDone := make(chan struct{})
@@ -168,7 +169,7 @@ func NewService(cfg *Config) (*Service, error) {
 
 		authData, err := srv.retrieveAuthorityData()
 		if err != nil {
-			return nil, err
+			return nil, fmt.Errorf("could not retrieve authority data: %s", err)
 		}
 
 		// BABE session configuration
@@ -270,7 +271,7 @@ func (s *Service) Stop() error {
 // StorageRoot returns the hash of the runtime storage root
 func (s *Service) StorageRoot() (common.Hash, error) {
 	if s.storageState == nil {
-		return common.Hash{}, ErrNilStorageState
+		return common.Hash{}, fmt.Errorf("storage state is nil")
 	}
 	return s.storageState.StorageRoot()
 }
@@ -289,7 +290,7 @@ func (s *Service) safeMsgSend(msg network.Message) error {
 	s.lock.Lock()
 	defer s.lock.Unlock()
 	if s.closed {
-		return ErrServiceStopped
+		return errors.New("service has been stopped")
 	}
 	s.msgSend <- msg
 	return nil
@@ -299,7 +300,7 @@ func (s *Service) safeBabeKill() error {
 	s.lock.Lock()
 	defer s.lock.Unlock()
 	if s.closed {
-		return ErrServiceStopped
+		return errors.New("service has been stopped")
 	}
 	close(s.babeKill)
 	return nil
@@ -398,7 +399,7 @@ func (s *Service) receiveMessages() {
 // handleReceivedBlock handles blocks from the BABE session
 func (s *Service) handleReceivedBlock(block *types.Block) (err error) {
 	if s.blockState == nil {
-		return ErrNilBlockState
+		return fmt.Errorf("blockState is nil")
 	}
 
 	err = s.blockState.AddBlock(block)
@@ -441,33 +442,33 @@ func (s *Service) handleReceivedMessage(msg network.Message) (err error) {
 	case network.BlockAnnounceMsgType:
 		blockAnnounceMessage, ok := msg.(*network.BlockAnnounceMessage)
 		if !ok {
-			return ErrMessageCast("BlockAnnounceMessage")
+			return errors.New("could not cast network.Message to BlockAnnounceMessage")
 		}
 
 		err = s.ProcessBlockAnnounceMessage(blockAnnounceMessage)
 	case network.BlockRequestMsgType:
 		msg, ok := msg.(*network.BlockRequestMessage)
 		if !ok {
-			return ErrMessageCast("BlockRequestMessage")
+			return errors.New("could not cast network.Message to BlockRequestMessage")
 		}
 
 		err = s.ProcessBlockRequestMessage(msg)
 	case network.BlockResponseMsgType:
 		msg, ok := msg.(*network.BlockResponseMessage)
 		if !ok {
-			return ErrMessageCast("BlockResponseMessage")
+			return errors.New("could not cast network.Message to BlockResponseMessage")
 		}
 
 		err = s.ProcessBlockResponseMessage(msg)
 	case network.TransactionMsgType:
 		msg, ok := msg.(*network.TransactionMessage)
 		if !ok {
-			return ErrMessageCast("TransactionMessage")
+			return errors.New("could not cast network.Message to TransactionMessage")
 		}
 
 		err = s.ProcessTransactionMessage(msg)
 	default:
-		err = ErrUnsupportedMsgType(msgType)
+		err = fmt.Errorf("Received unsupported message type %d", msgType)
 	}
 
 	return err
@@ -485,19 +486,19 @@ func (s *Service) ProcessBlockAnnounceMessage(msg *network.BlockAnnounceMessage)
 	}
 
 	_, err = s.blockState.GetHeader(header.Hash())
-	if err != nil && err == database.ErrKeyNotFound {
+	if err != nil && err.Error() == "Key not found" {
 		err = s.blockState.SetHeader(header)
 		if err != nil {
 			return err
 		}
 
 		log.Info("[core] saved block", "number", header.Number, "hash", header.Hash())
-	} else if err != nil {
+	} else {
 		return err
 	}
 
 	_, err = s.blockState.GetBlockBody(header.Hash())
-	if err != nil && err == database.ErrKeyNotFound {
+	if err != nil && err.Error() == "Key not found" {
 		// send block request message
 		log.Debug("[core] sending new block to syncer", "number", msg.Number)
 		s.blockNumOut <- msg.Number
@@ -642,7 +643,7 @@ func (s *Service) checkForRuntimeChanges() error {
 
 		s.rt.Stop()
 
-		s.rt, err = runtime.NewRuntime(code, s.storageState, s.keys)
+		s.rt, err = runtime.NewRuntime(code, s.storageState, s.keys, runtime.RegisterImports)
 		if err != nil {
 			return err
 		}
