+++ lib/runtime/runtime.go
@@ -23,17 +23,18 @@ import (
 	"sync"
 
 	"github.com/ChainSafe/gossamer/lib/keystore"
-	allocator "github.com/ChainSafe/gossamer/lib/runtime/allocator"
 	"github.com/ChainSafe/gossamer/lib/scale"
 
 	log "github.com/ChainSafe/log15"
 	wasm "github.com/wasmerio/go-ext-wasm/wasmer"
 )
 
+var memory, memErr = wasm.NewMemory(17, 0)
+
 // Ctx struct
 type Ctx struct {
 	storage   Storage
-	allocator *allocator.FreeingBumpHeapAllocator
+	allocator *FreeingBumpHeapAllocator
 	keystore  *keystore.Keystore
 }
 
@@ -43,22 +44,22 @@ type Runtime struct {
 	storage   Storage
 	keystore  *keystore.Keystore
 	mutex     sync.Mutex
-	allocator *allocator.FreeingBumpHeapAllocator
+	allocator *FreeingBumpHeapAllocator
 }
 
 // NewRuntimeFromFile instantiates a runtime from a .wasm file
-func NewRuntimeFromFile(fp string, s Storage, ks *keystore.Keystore) (*Runtime, error) {
+func NewRuntimeFromFile(fp string, s Storage, ks *keystore.Keystore, registerImports func() (*wasm.Imports, error)) (*Runtime, error) {
 	// Reads the WebAssembly module as bytes.
 	bytes, err := wasm.ReadBytes(fp)
 	if err != nil {
 		return nil, err
 	}
 
-	return NewRuntime(bytes, s, ks)
+	return NewRuntime(bytes, s, ks, registerImports)
 }
 
 // NewRuntime instantiates a runtime from raw wasm bytecode
-func NewRuntime(code []byte, s Storage, ks *keystore.Keystore) (*Runtime, error) {
+func NewRuntime(code []byte, s Storage, ks *keystore.Keystore, registerImports func() (*wasm.Imports, error)) (*Runtime, error) {
 	if s == nil {
 		return nil, errors.New("runtime does not have storage trie")
 	}
@@ -74,7 +75,15 @@ func NewRuntime(code []byte, s Storage, ks *keystore.Keystore) (*Runtime, error)
 		return nil, err
 	}
 
-	memAllocator := allocator.NewAllocator(instance.Memory, 0)
+	if instance.Memory == nil {
+		if memErr != nil {
+			return nil, err
+		}
+
+		instance.Memory = memory
+	}
+
+	memAllocator := NewAllocator(instance.Memory, 0)
 
 	runtimeCtx := Ctx{
 		storage:   s,
