--- lib/runtime/wasmer_test_func.go
@@ -1,52 +1,0 @@
-// Copyright 2019 ChainSafe Systems (ON) Corp.
-// This file is part of gossamer.
-//
-// The gossamer library is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Lesser General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// The gossamer library is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU Lesser General Public License for more details.
-//
-// You should have received a copy of the GNU Lesser General Public License
-// along with the gossamer library. If not, see <http://www.gnu.org/licenses/>.
-
-package runtime
-
-import (
-	"path/filepath"
-	"testing"
-
-	"github.com/ChainSafe/gossamer/lib/keystore"
-	"github.com/ChainSafe/gossamer/lib/trie"
-	"github.com/ChainSafe/gossamer/tests"
-
-	"github.com/stretchr/testify/require"
-)
-
-// NewTestRuntime will create a new runtime (polkadot/test)
-func NewTestRuntime(t *testing.T, targetRuntime string) *Runtime {
-	return NewTestRuntimeWithTrie(t, targetRuntime, nil)
-}
-
-// NewTestRuntimeWithTrie will create a new runtime (polkadot/test) with the supplied trie as the storage
-func NewTestRuntimeWithTrie(t *testing.T, targetRuntime string, tt *trie.Trie) *Runtime {
-	testRuntimeFilePath, testRuntimeURL := tests.GetRuntimeVars(targetRuntime)
-
-	_, err := tests.GetRuntimeBlob(testRuntimeFilePath, testRuntimeURL)
-	require.Nil(t, err, "Fail: could not get runtime", "targetRuntime", targetRuntime)
-
-	rs := tests.NewTestRuntimeStorage(tt)
-
-	fp, err := filepath.Abs(testRuntimeFilePath)
-	require.Nil(t, err, "could not create testRuntimeFilePath", "targetRuntime", targetRuntime)
-
-	r, err := NewRuntimeFromFile(fp, rs, keystore.NewKeystore())
-	require.Nil(t, err, "Got error when trying to create new VM", "targetRuntime", targetRuntime)
-	require.NotNil(t, r, "Could not create new VM instance", "targetRuntime", targetRuntime)
-
-	return r
-}
